Combine AbstractReadHandler and BlockReadHandler
